+++ src/pages/networking-create/services/networking.ts
@@ -1,0 +1,36 @@
+import axiosInstance from '@/api/axiosInstance';
+
+export interface CreateNetworkingRequestBody {
+  title: string;
+  contents: string;
+  maxNumber: number;
+  representativeId: number;
+}
+
+export interface CreateNetworkingResponseBody {
+  id: number;
+  title: string;
+  contents: string;
+  maxNumber: number;
+  representativeId: number;
+}
+
+export interface CreateNetworkingParams {
+  requestBody: CreateNetworkingRequestBody;
+  chatRoomId?: number;
+}
+
+export async function createNetworking({
+  requestBody,
+  chatRoomId,
+}: CreateNetworkingParams): Promise<CreateNetworkingResponseBody> {
+  const { data } = await axiosInstance.post<CreateNetworkingResponseBody>(
+    '/networking',
+    requestBody,
+    {
+      params: chatRoomId ? { chat_room_id: chatRoomId } : undefined,
+    },
+  );
+
+  return data;
+}
